requirements.txt
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.4.26
colorama==0.4.6
Django==5.2.1
django-extensions==4.1
djangorestframework==3.16.0
h11==0.16.0
holehe==1.61
httpcore==1.0.9
httpx==0.28.1
idna==3.10
outcome==1.3.0.post0
pydot==4.0.0
pygraphviz==1.14
pyparsing==3.2.3
setuptools==80.8.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
termcolor==3.1.0
tqdm==4.67.1
trio==0.30.0
typing_extensions==4.13.2
